task: from review comment, move filterCard to separate section
