+++ packages/components-library/src/main.ts
@@ -2,9 +2,10 @@ import CartSelectionToast from './components/ui/CartSelectionToast.vue'
 import CheckboxFilter from './components/filters/CheckboxFilter.vue'
 import StringFilter from './components/filters/StringFilter.vue'
 import NumberFilter from './components/filters/NumberFilter.vue'
+import FilterCard from './components/filters/containers/FilterCard.vue'
 
-export { CartSelectionToast, CheckboxFilter, StringFilter, NumberFilter }
+export { CartSelectionToast, CheckboxFilter, StringFilter, NumberFilter, FilterCard }
 
 export default {
-  CartSelectionToast, CheckboxFilter, StringFilter, NumberFilter
+  CartSelectionToast, CheckboxFilter, StringFilter, NumberFilter, FilterCard
 }
